requirements.txt
aiogram==3.20.0.post0
yookassa
requests
python-dotenv
aiogram
python-dotenv
...